requirements.txt
Flask>=3.1.0
Flask-Login>=0.6.3
Flask-SocketIO>=5.6.0
python-socketio>=5.16.0
Werkzeug>=3.1.0
resend>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
eventlet>=0.33.3
Flask-WTF>=1.2.1
Flask-Limiter>=3.5.0
email-validator>=2.1.0
pytz
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
reportlab>=4.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
